requirements.txt
replicate~=1.0.4
openai~=1.61.1
pytest-playwright~=0.7.0
pillow~=11.1.0
playwright~=1.50.0
PyQt6~=6.9.0
diffusers~=0.34.0
transformers~=4.53.1
accelerate~=1.8.1
safetensors~=0.5.3
torch~=2.7.1
# compel and torchvision are optional for basic diffusers, exclude due to space
# compel~=2.1.1
# torchvision~=0.22.1
ollama~=0.5.1
requests~=2.32.0